Refactor. 코드리뷰사항 반영
- 매직 넘버 제거, 변수에 할당
- Rolling > render 메서드 reduce -> map & join으로 수정
- util delay 함수 parameter 순서 변경
- Rolling > roll 메서드 setTimeout 없애고 delay로 통일
