+++ public/js/main.js
@@ -1,7 +1,7 @@
 import { _, getData } from './util/util.js';
 import { $searchBar, $searchBarInput, $searchSuggestions, $rollingKeywords } from './util/ref.js';
 import { SearchBar, Rolling, KeywordSuggestion, AutoComplete } from './searchBar.js';
-import { URL, KEYCODE } from './util/data.js';
+import { URL, KEYCODE } from './util/constants.js';
 import { parseRecommendedList } from './util/parser.js';
 const { log } = console;
 
@@ -18,10 +18,8 @@ searchBar.registerEvent();
 
 // rolling & search keyword suggestion
 async function initSuggestion() {
-  const recommendJsonData = await getData(URL.RECOMMEND);
-  
   // rolling
-  const { list: recommendListInfo } = recommendJsonData;
+  const { list: recommendListInfo } = await getData(URL.RECOMMEND);
   const recommendList = parseRecommendedList(recommendListInfo); // 키워드만 뽑아서 10개로 자름
   const rollingArgs = {
     domElem: $rollingKeywords,
