+++ public/js/searchBar.js
@@ -1,6 +1,6 @@
 import { CLASS_LIST } from './util/cssClasses.js';
-import { delay } from './util/util.js';
-import { URL, KEYCODE } from './util/data.js';
+import { delay, getData } from './util/util.js';
+import { URL, KEYCODE } from './util/constants.js';
 
 class SearchBar {
   constructor({ domElem, inputArea, suggestions, rollingKeywords }) {
@@ -30,33 +30,44 @@ class SearchBar {
 }
 
 class Rolling {
+  static MAX_ITEM_COUNT = 10;
+
   constructor({ domElem, rollingData }) {
     this.domElem = domElem;
     this.rollingData = rollingData;
+    this.initialY = 4;
     this.currItemIdx = 0;
-    this.currY = 4;
+    this.currY = this.initialY;
     this.heightOfOneItem = 30;
     this.init();
   }
 
   render() {
-    let renderingValue = this.rollingData.reduce((prev, curr, idx) => {
-      return prev + `<li><span class="num-rank">${idx + 1}</span>${this.rollingData[idx]}</li>`
-    }, '')
+    let renderingValue = this.rollingData.map((keyword, idx) => `<li><span class="num-rank">${idx + 1}</span>${keyword}</li>`).join('');
     renderingValue += `<li><span class="num-rank">1</span>${this.rollingData[0]}</li>`
     this.domElem.innerHTML = renderingValue;
   }
 
   async roll() {
-    await delay('', 2000);
-    const rollngTimeoutId = setTimeout(() => {
-      this.setAnimation();
-      this.currItemIdx++;
-      if(this.currItemIdx === 11) {
-        this.resetStates();
-      }
-      this.roll()
-    }, 1000);
+    // 👇 나중에 timeout을 정지시켜야 할 일이 있어서 setTimeout 남겨둠
+    // await delay(2000);
+    // const rollngTimeoutId = setTimeout(() => {
+    //   this.setAnimation();
+    //   this.currItemIdx++;
+    //   if(this.currItemIdx === 11) {
+    //     this.resetStates();
+    //   }
+    //   this.roll()
+    // }, 1000);
+
+    // 👇 setTimeout을 delay로 바꾼 버전
+    await delay(3000);
+    this.setAnimation();
+    this.currItemIdx++;
+    if(this.currItemIdx === Rolling.MAX_ITEM_COUNT + 1) {
+      this.resetStates();
+    }
+    this.roll();
   }
 
   setAnimation() {
@@ -66,13 +77,13 @@ class Rolling {
   }
 
   resetStates() {
-    this.currY = 4;
+    this.currY = this.initialY;
     this.domElem.style.transform = '';
     this.domElem.style.transition = '';
     this.currItemIdx = 0;
   }
 
-  async init(){
+  init() {
     this.render();
     this.roll();
   }
