+++ public/js/util/util.js
@@ -6,7 +6,7 @@ const _ = {
 const pipe = (...fns) => arg => fns.reduce((arg, fn) => fn(arg), arg);
 const createDom = (tag) => ({ value, classes }) => `<${tag} class='${classes.join(' ')}'>${value}</${tag}>`;
 const getData = (url) => fetch(url).then((res) => res.json());
-const delay = (value, ms) => new Promise((resolve) => setTimeout(() => resolve(value), ms));
+const delay = (ms, value = '') => new Promise((resolve) => setTimeout(() => resolve(value), ms));
 const debounce = (fn, ms) => {
   let timeoutID;
   return function(...args) {
